@@ -1,6 +1,6 @@
 module github.com/deatil/lakego-jwt
 
-go 1.16
+go 1.18
 
 require (
 	github.com/golang-jwt/jwt/v4 v4.4.1
